requirements.txt
scikit_image
git+https://gitlab.com/jixingcn/blender-mathutils.git
git+https://github.com/scaleapi/pandaset-devkit.git#subdirectory=python
tqdm
pandas
matplotlib
scipy
pyquaternion
kornia
Pillow
wandb
pyntcloud
git+https://github.com/utiasSTARS/liegroups
geomstats
torch_geometric==2.0.0
yacs
timm==0.4.12
einops
loguru
open3d
rich
# For argoverse:
omegaconf==2.1.0
colour
descartes
h5py
hydra-core==1.1.0
lapsolver
motmetrics==1.1.3
numba
polars
shapely
scikit_learn
